+++ oshdb/src/main/java/org/heigit/bigspatialdata/oshdb/osh/OSHNode.java
@@ -11,7 +11,7 @@ import org.heigit.bigspatialdata.oshdb.osh.builder.Builder;
 import org.heigit.bigspatialdata.oshdb.osm.OSMEntity;
 import org.heigit.bigspatialdata.oshdb.osm.OSMNode;
 import org.heigit.bigspatialdata.oshdb.osm.OSMType;
-import org.heigit.bigspatialdata.oshdb.util.BoundingBox;
+import org.heigit.bigspatialdata.oshdb.util.OSHDBBoundingBox;
 import org.heigit.bigspatialdata.oshdb.util.byteArray.ByteArrayOutputWrapper;
 import org.heigit.bigspatialdata.oshdb.util.byteArray.ByteArrayWrapper;
 
@@ -38,14 +38,14 @@ public class OSHNode extends OSHEntity<OSMNode> implements Iterable<OSMNode>, Se
     ByteArrayWrapper wrapper = ByteArrayWrapper.newInstance(data, offset, length);
     //header holds data on bitlevel and can then be compared to stereotypical bitcombinations (e.g. this.HEADER_HAS_TAGS)
     final byte header = wrapper.readRawByte();
-    final BoundingBox bbox;
+    final OSHDBBoundingBox bbox;
     if ((header & HEADER_HAS_BOUNDINGBOX) != 0) {
       final long minLon = baseLongitude + wrapper.readSInt64();
       final long maxLon = minLon + wrapper.readUInt64();
       final long minLat = baseLatitude + wrapper.readSInt64();
       final long maxLat = minLat + wrapper.readUInt64();
 
-      bbox = new BoundingBox(minLon, maxLon, minLat, maxLat);
+      bbox = new OSHDBBoundingBox(minLon, minLat, maxLon, maxLat);
 
     } else {
       bbox = null;
@@ -73,7 +73,7 @@ public class OSHNode extends OSHEntity<OSMNode> implements Iterable<OSMNode>, Se
 
   private OSHNode(final byte[] data, final int offset, final int length, final long baseNodeId,
           final long baseTimestamp, final long baseLongitude, final long baseLatitude, final byte header,
-          final long id, final BoundingBox bbox, final int[] keys, final int dataOffset, final int dataLength) {
+          final long id, final OSHDBBoundingBox bbox, final int[] keys, final int dataOffset, final int dataLength) {
     super(data, offset, length, baseNodeId, baseTimestamp, baseLongitude, baseLatitude, header, id, bbox, keys,
             dataOffset, dataLength);
   }
@@ -90,7 +90,7 @@ public class OSHNode extends OSHEntity<OSMNode> implements Iterable<OSMNode>, Se
   }
 
   @Override
-  public BoundingBox getBoundingBox() {
+  public OSHDBBoundingBox getBoundingBox() {
     if (bbox != null) {
       return bbox;
     }
@@ -113,7 +113,7 @@ public class OSHNode extends OSHEntity<OSMNode> implements Iterable<OSMNode>, Se
       return null;
     }
 
-    return new BoundingBox(minLon, maxLon, minLat, maxLat);
+    return new OSHDBBoundingBox(minLon, minLat, maxLon, maxLat);
   }
 
   @Override