+++ oshdb/src/test/java/org/heigit/bigspatialdata/oshdb/util/OSHDBBoundingBoxTest.java
@@ -7,14 +7,14 @@ import static org.junit.Assert.assertArrayEquals;
 import static org.junit.Assert.assertEquals;
 import org.junit.Test;
 
-public class BoundingBoxTest {
+public class OSHDBBoundingBoxTest {
 
-  public BoundingBoxTest() {
+  public OSHDBBoundingBoxTest() {
   }
 
   @Test
   public void testToString() {
-    BoundingBox instance = new BoundingBox(0.0, 1.0, 89.0, 90.0);
+    OSHDBBoundingBox instance = new OSHDBBoundingBox(0.0, 1.0, 89.0, 90.0);
     String expResult = "(0.000000,1.000000) (89.000000,90.000000)";
     String result = instance.toString();
     assertEquals(expResult, result);
@@ -22,25 +22,25 @@ public class BoundingBoxTest {
 
   @Test
   public void testIntersect() {
-    BoundingBox first = new BoundingBox(0.0, 1.0, 89.0, 90.0);
-    BoundingBox second = new BoundingBox(0.0, 2.0, 89.9, 90.0);
-    BoundingBox expResult = new BoundingBox(0.0, 1.0, 89.9, 90.0);
-    BoundingBox result = BoundingBox.intersect(first, second);
+    OSHDBBoundingBox first = new OSHDBBoundingBox(0.0, 89.0, 1.0, 90.0);
+    OSHDBBoundingBox second = new OSHDBBoundingBox(0.0, 89.9, 2.0, 90.0);
+    OSHDBBoundingBox expResult = new OSHDBBoundingBox(0.0, 89.9, 1.0, 90.0);
+    OSHDBBoundingBox result = OSHDBBoundingBox.intersect(first, second);
     assertEquals(expResult, result);
   }
 
   @Test
   public void testOverlap() {
-    BoundingBox a = new BoundingBox(0.1, 0.9, 89.1, 89.9);
-    BoundingBox b = new BoundingBox(0.0, 1.0, 89.0, 90.0);
-    BoundingBox.OVERLAP expResult = BoundingBox.OVERLAP.A_COMPLETE_IN_B;
-    BoundingBox.OVERLAP result = BoundingBox.overlap(a, b);
+    OSHDBBoundingBox a = new OSHDBBoundingBox(0.1, 89.1, 0.9, 89.9);
+    OSHDBBoundingBox b = new OSHDBBoundingBox(0.0, 89.0, 1.0, 90.0);
+    OSHDBBoundingBox.OVERLAP expResult = OSHDBBoundingBox.OVERLAP.A_COMPLETE_IN_B;
+    OSHDBBoundingBox.OVERLAP result = OSHDBBoundingBox.overlap(a, b);
     assertEquals(expResult, result);
   }
 
   @Test
   public void testGetMinLon() {
-    BoundingBox instance = new BoundingBox(0.0, 1.0, 89.0, 90.0);
+    OSHDBBoundingBox instance = new OSHDBBoundingBox(0.0, 89.0, 1.0, 90.0);
     double expResult = 0.0;
     double result = instance.getMinLon();
     assertEquals(expResult, result, 0.0);
@@ -48,7 +48,7 @@ public class BoundingBoxTest {
 
   @Test
   public void testGetMaxLon() {
-    BoundingBox instance = new BoundingBox(0.0, 1.0, 89.0, 90.0);
+    OSHDBBoundingBox instance = new OSHDBBoundingBox(0.0, 89.0, 1.0, 90.0);
     double expResult = 1.0;
     double result = instance.getMaxLon();
     assertEquals(expResult, result, 0.0);
@@ -56,7 +56,7 @@ public class BoundingBoxTest {
 
   @Test
   public void testGetMinLat() {
-    BoundingBox instance = new BoundingBox(0.0, 1.0, 89.0, 90.0);
+    OSHDBBoundingBox instance = new OSHDBBoundingBox(0.0, 89.0, 1.0, 90.0);
     double expResult = 89.0;
     double result = instance.getMinLat();
     assertEquals(expResult, result, 0.0);
@@ -64,7 +64,7 @@ public class BoundingBoxTest {
 
   @Test
   public void testGetMaxLat() {
-    BoundingBox instance = new BoundingBox(0.0, 1.0, 89.0, 90.0);
+    OSHDBBoundingBox instance = new OSHDBBoundingBox(0.0, 89.0, 1.0, 90.0);
     double expResult = 90.0;
     double result = instance.getMaxLat();
     assertEquals(expResult, result, 0.0);
@@ -72,7 +72,7 @@ public class BoundingBoxTest {
 
   @Test
   public void testGetLon() {
-    BoundingBox instance = new BoundingBox(0.0, 1.0, 89.0, 90.0);
+    OSHDBBoundingBox instance = new OSHDBBoundingBox(0.0, 89.0, 1.0, 90.0);
     long[] expResult = new long[]{0L, 10000000L};
     long[] result = instance.getLon();
     assertArrayEquals(expResult, result);
@@ -80,7 +80,7 @@ public class BoundingBoxTest {
 
   @Test
   public void testGetLat() {
-    BoundingBox instance = new BoundingBox(0.0, 1.0, 89.0, 90.0);
+    OSHDBBoundingBox instance = new OSHDBBoundingBox(0.0, 89.0, 1.0, 90.0);
     long[] expResult = new long[]{890000000L, 900000000L};
     long[] result = instance.getLat();
     assertArrayEquals(expResult, result);
@@ -88,7 +88,7 @@ public class BoundingBoxTest {
 
   @Test
   public void testGetGeometry() {
-    BoundingBox instance = new BoundingBox(0.0, 1.0, 0.0, 1.0);
+    OSHDBBoundingBox instance = new OSHDBBoundingBox(0.0, 0.0, 1.0, 1.0);
     Polygon expResult = (new GeometryFactory()).createPolygon(new Coordinate[]{new Coordinate(0, 0), new Coordinate(1, 0), new Coordinate(1, 1), new Coordinate(0, 1), new Coordinate(0, 0)});
     Polygon result = instance.getGeometry();
     assertEquals(expResult, result);
@@ -96,7 +96,7 @@ public class BoundingBoxTest {
 
   @Test
   public void testHashCode() {
-    BoundingBox instance = new BoundingBox(0.0, 1.0, 89.0, 90.0);
+    OSHDBBoundingBox instance = new OSHDBBoundingBox(0.0, 89.0, 1.0, 90.0);
     int expResult = 748664391;
     int result = instance.hashCode();
     assertEquals(expResult, result);
@@ -104,8 +104,8 @@ public class BoundingBoxTest {
 
   @Test
   public void testEquals() {
-    Object obj = new BoundingBox(0.0, 1.0, 89.0, 90.0);
-    BoundingBox instance = new BoundingBox(0.0, 1.0, 89.0, 90.0);
+    Object obj = new OSHDBBoundingBox(0.0, 89.0, 1.0, 90.0);
+    OSHDBBoundingBox instance = new OSHDBBoundingBox(0.0, 89.0, 1.0, 90.0);
     boolean expResult = true;
     boolean result = instance.equals(obj);
     assertEquals(expResult, result);